Fix in User Class

+++ classes/User.class.php
@@ -15,7 +15,7 @@ class User {
     private $_last_login;
     private $_avatar;
     private $_badges = array();
-	private $_courses = array();
+    private $_courses = array();
 
     public function __construct($db, $id) {
         $this->_db = $db;
@@ -23,7 +23,7 @@ class User {
 
         $result = $db->select("users", "*", array("id" => $this->_id));
         $badges = $this->_db->select("users_badges", array("badge_id"), array("user_id" => $this->_id));
-		$courses = $this->_db->select("users_courses", array("course_id"), array("user_id" => $this->_id));
+        $courses = $this->_db->select("users_courses", array("course_id"), array("user_id" => $this->_id));
 
         $this->setBirthDate($result[0]["birth_date"]);
         $this->setEmail($result[0]["email"]);
@@ -41,8 +41,8 @@ class User {
                 $this->_badges[] = $value['badge_id'];
             }
         }
-		
-		if ($courses != null) {
+
+        if ($courses != null) {
             foreach ($courses as $value) {
                 $this->_courses[] = $value['course_id'];
             }
@@ -166,6 +166,18 @@ class User {
      */
 
     public function delete() {
+        $result = $this->_db->delete("users_badges", array("user_id" => $this->_id));
+        if ($result == false) {
+            return false;
+        }
+        $result = $this->_db->delete("users_clusters", array("user_id" => $this->_id));
+        if ($result == false) {
+            return false;
+        }
+        $result = $this->_db->delete("users_courses", array("user_id" => $this->_id));
+        if ($result == false) {
+            return false;
+        }
         $result = $this->_db->delete("users", array("id" => $this->_id));
         if ($result == false) {
             return false;
@@ -271,8 +283,8 @@ class User {
         }
         return $badges;
     }
-	
-	/*
+
+    /*
      * getCourses()
      * 
      * Verkrijg de lessen waarin een gebruiker voortgang heeft geboekt.
@@ -287,19 +299,19 @@ class User {
         }
         return $courses;
     }
-	
-	/*
+
+    /*
      * setCourseXP()
      * 
      * Verkrijg de lessen waarin een gebruiker voortgang heeft geboekt.
      * 
-	 * @param Course $course De les waarvan de XP moet worden ingesteld.
+     * @param Course $course De les waarvan de XP moet worden ingesteld.
      * @result Boolean Succesvol of niet.
      */
-	
-	public function setCourseXP($course,$xp) {
-		$this->_db->update("courses",array("xp_earned"=>$xp),array("user_id"=>$this->_id,"course_id"=>$course->getID()));
-	}
+
+    public function setCourseXP($course, $xp) {
+        $this->_db->update("courses", array("xp_earned" => $xp), array("user_id" => $this->_id, "course_id" => $course->getID()));
+    }
 
     /*
      * getID()
@@ -443,7 +455,7 @@ class User {
     public function getLastLogin() {
         return $this->_last_login;
     }
-    
+
     public function getPassword() {
         $result = $this->_db->select("users", array("password"), array("id" => $this->_id));
         return $result[0]['password'];
